refactor(storage): collapse unwrap clauses, share CountingCatalog, drop the iterator's eager table load
